Pression

+++ config_cryst_3D.c
@@ -9,7 +9,7 @@ int main (int argc, char *argv[]) {
         sscanf(argv[1],"%d",&n);
         sscanf(argv[2],"%lf",&box_size);
     }
-    int dx = box_size/n; 
+    double dx = box_size/n; 
     posx = (-box_size+dx)/2;
     FILE *file;
